Read configuration at Startup
When a Quay token is rotated, the new values
are projected into the /workspace folder.
The controller is loading the secret data
only at startup, so it won't see updates.
This implies that to effectively rotate the
Quay token you also need to restart the controller.
With this patch, pod restart should not be required.

Signed-off-by: Francesco Ilario <filario@redhat.com>