Temporarily disable image pruner (#115)
STONEBLD-2466

The pruner may improperly delete the 'sha256-${digest}.*' attachments
for images that still exist.

The pruner is not supposed to delete these attachments if the
'sha256:${digest}' image exists. However, it only takes into account
tagged images. The individual image manifests (e.g. the arch-specific
images) do not have to be tagged, it is enough if they are part of an
image index (e.g. the multi-arch image).

Disable the pruner until the issue is fixed properly.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>