[ISV-5681] Cronjob to reset repository notifications (#183)
This cronjob is intended to regularly (once in a day initially) iterate over all Quay repositories in a given namespace and reset all notifications failure counter to 0 if needed.
The job will reset counter for all notifications regardless if it is a push webhook defined within ImageRepository (the original reason for this cronjob to exist) or manually defined by repo user/owner. Making the job aware of the ImageRepository definitions was considered as a significant overhead for it's purpose.